Use MarkdownIt singleton via svelte context
Fixes #103
